# db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.16.0

package ttssqlc

import (
	"database/sql"
	"time"
)

type SchemaMigration struct {
	Version int64
	Dirty   bool
}

type Session struct {
	ID        int64
	UserID    int64
	Token     string
	CreatedAt time.Time
}

type Task struct {
	ID          int64
	Title       string
	Description string
	Result      string
	CreatedAt   time.Time
	CreatedBy   int64
	DoneAt      sql.NullTime
}

type TasksDependency struct {
	ID          int64
	TaskID      int64
	DependsOnID int64
	CreatedAt   time.Time
}

type User struct {
	ID         int64
	Username   string
	Password   sql.NullString
	Email      sql.NullString
	Confirmed  sql.NullBool
	CreatedAt  time.Time
	RootTaskID sql.NullInt64
}